+++ gocd/gocd.go
@@ -73,6 +73,7 @@ type Client struct {
 	PipelineConfigs   *PipelineConfigsService
 	Configuration     *ConfigurationService
 	Encryption        *EncryptionService
+	Plugins           *PluginsService
 
 	common service
 	cookie string
@@ -152,6 +153,7 @@ func NewClient(cfg *Configuration, httpClient *http.Client) *Client {
 	c.PipelineConfigs = (*PipelineConfigsService)(&c.common)
 	c.Configuration = (*ConfigurationService)(&c.common)
 	c.Encryption = (*EncryptionService)(&c.common)
+	c.Plugins = (*PluginsService)(&c.common)
 
 	return c
 }