+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Expanded test cases for jobs and tasks
  - Added Pipeline status querying and pipeline history structure
  - Added Basic get-configuration command.
+ - Added plugin retrieval
  
 ### Changed
  - Fixed a bug where if the config file was not found the cli tool would crash.